Created an AddRemoveFriends component that will be called into FriendsContainer.
